requirements.txt
amqp==2.5.2
apispec==1.3.3
appdirs==1.4.4
astroid==2.4.1
attrs==19.3.0
Babel==2.8.0
billiard==3.6.3.0
black==19.10b0
blinker==1.4
boto3==1.13.16
botocore==1.16.16
celery==4.4.2
click==7.1.2
colorama==0.4.3
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
Flask==1.1.2
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Celery==2.4.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
flower==0.9.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.0
mccabe==0.6.1
pathspec==0.8.0
Pillow==7.2.0
prison==0.1.3
PyJWT==1.7.1
pylint==2.5.2
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-slugify==4.0.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
regex==2020.5.14
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.5
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
waitress==1.4.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
